+++ CHANGELOG.md
@@ -1,8 +1,20 @@
 # Changelog
 
-## 0.1.0 (2015-11-11)
+## v0.1.0 (2015-11-11)
 
 Initial public release
 
 - Add documentation
 - Add unit tests
+
+## v0.2.0 (2016-12-15)
+
+**Implemented enhancements:**
+- Add unix socket support [\#4](https://github.com/aristanetworks/goeapi/issues/4)
+- Add support for Connect Timeout [\#2]( https://github.com/aristanetworks/goeapi/issues/2)
+
+
+**Fixed Bugs:**
+- Bump up travis to go1.7 [\#10](https://github.com/aristanetworks/goeapi/issues/10)
+- Url endpoint shouldn't terminate in '/' [\#8]( https://github.com/aristanetworks/goeapi/issues/8)
+- Fix examples and vet issues [\#6]( https://github.com/aristanetworks/goeapi/issues/6)